# core/engines/genomic-foundation/web-portal/openapi.yaml
openapi: "3.0.3"
info:
  title: Genomics Pipeline Web Portal API
  description: |
    REST API for the HCLS AI Factory Genomics Pipeline Web Portal.
    Manages the FASTQ to VCF workflow powered by NVIDIA Clara Parabricks
    on DGX Spark hardware. Provides pipeline execution, monitoring,
    file management, and real-time system metrics.
  version: "1.0.0"
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:5000
    description: Local development server
security: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Required for mutating endpoints when the PORTAL_API_KEY environment
        variable is set. When PORTAL_API_KEY is unset, authentication is
        disabled (local dev mode).
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    UploadResponse:
      type: object
      properties:
        success:
          type: boolean
        filename:
          type: string
        size:
          type: integer
          format: int64
        size_human:
          type: string
...
  responses:
...
    RateLimited:
      description: Rate limit exceeded.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              retry_after:
                type: integer
          example:
            error: Rate limit exceeded
            retry_after: 60
...
  parameters:
    DirectoryParam:
      name: directory
      in: path
      required: true
      schema:
        type: string
        enum: [input, output, ref]
      description: Data subdirectory name.
...
paths:
...
  /api/files/upload/{directory}:
    post:
      operationId: uploadFile
      summary: Upload a file
      description: |
        Uploads a file to the specified data directory. Rate limited.
        Maximum upload size is 100 GB.
      tags: [Files]
      parameters:
        - $ref: "#/components/parameters/DirectoryParam"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: File uploaded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
        "400":
          description: Invalid directory, missing file, or invalid filename.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          description: Error saving file.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...